% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{csvdata}
\alias{csvdata}
\title{csvdata' 'data_to_json' 'drawr' 'linearDataFileGen' 'linearDataGen'
linearModel' 'load_file' 'nonLinearDataFileGen' 'nonLinearDataGen'
nonLinearModel' 'ydtApp' 'ydtServer' 'ydtUI'
stores csvdata imported from file
csvdata}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
csvdata
}
\description{
csvdata' 'data_to_json' 'drawr' 'linearDataFileGen' 'linearDataGen'
linearModel' 'load_file' 'nonLinearDataFileGen' 'nonLinearDataGen'
nonLinearModel' 'ydtApp' 'ydtServer' 'ydtUI'
stores csvdata imported from file
csvdata
}
\keyword{datasets}
