% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{drawr}
\alias{drawr}
\title{drawr}
\usage{
drawr(
  data,
  linear = "true",
  draw_start = NULL,
  points_end = NULL,
  x_by = 0.25,
  free_draw = T,
  points = "partial",
  aspect_ratio = 1.5,
  title = "",
  x_range = NULL,
  y_range = NULL,
  x_lab = "",
  y_lab = "",
  drawn_line_color = "steelblue",
  data_tab1_color = "steelblue",
  x_axis_buffer = 0.01,
  y_axis_buffer = 0.05,
  show_finished = T,
  shiny_message_loc = NULL
)
}
\arguments{
\item{data}{data}

\item{linear}{if linear model}

\item{draw_start}{start point}

\item{points_end}{end points}

\item{x_by}{x increment by}

\item{free_draw}{free draw}

\item{points}{points to draw}

\item{aspect_ratio}{aspect ratio}

\item{title}{title of chart}

\item{x_range}{range of x axis}

\item{y_range}{range of y axis}

\item{x_lab}{label for x}

\item{y_lab}{label for y}

\item{drawn_line_color}{color of line}

\item{data_tab1_color}{color of tab}

\item{x_axis_buffer}{buffer for x}

\item{y_axis_buffer}{buffer for y}

\item{show_finished}{show}

\item{shiny_message_loc}{message}
}
\description{
drawr
}
