% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{linearDataFileGen}
\alias{linearDataFileGen}
\title{linearDataFileGen}
\usage{
linearDataFileGen(
  y_xbar,
  slope,
  sigma,
  points_choice = "full",
  points_end_scale = 1,
  N = 30,
  x_min = 0,
  x_max = 20,
  x_by = 0.25
)
}
\arguments{
\item{y_xbar}{y_xbar}

\item{slope}{slope}

\item{sigma}{sigma}

\item{points_choice}{points_choice}

\item{points_end_scale}{points_choice}

\item{N}{number of points}

\item{x_min}{min x}

\item{x_max}{max x}

\item{x_by}{increment by}
}
\description{
linearDataFileGen
}
