% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{linearModel}
\alias{linearModel}
\title{stores linearModel for prediction
linearModel}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
linearModel
}
\description{
- linear model for fitting linear curve using least square regression
}
\keyword{datasets}
