% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{nonLinearModel}
\alias{nonLinearModel}
\title{stores nonLinearModel for prediction
nonLinearModel}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
nonLinearModel
}
\description{
nonLinearModel - non linear model for fitting no linear curve using least square regression
}
\keyword{datasets}
